+++ bika/lims/browser/analysisrequest/publish.py
@@ -26,6 +26,7 @@ from Products.CMFCore.WorkflowCore import WorkflowException
 from Products.CMFCore.utils import getToolByName
 from Products.CMFPlone.utils import _createObjectByType, safe_unicode
 from Products.Five.browser.pagetemplatefile import ViewPageTemplateFile
+from bika.lims import api
 from bika.lims import POINTS_OF_CAPTURE, bikaMessageFactory as _, t
 from bika.lims import logger
 from bika.lims.browser import BrowserView, ulocalized_time
@@ -36,6 +37,7 @@ from bika.lims.interfaces.field import IUIDReferenceField
 from bika.lims.utils import attachPdf, createPdf, encode_header, \
     format_supsub, \
     isnumber
+from bika.lims.utils import convert_unit
 from bika.lims.utils import formatDecimalMark, to_utf8
 from bika.lims.utils.analysis import format_uncertainty
 from bika.lims.vocabularies import getARReportTemplates
@@ -70,7 +72,8 @@ class AnalysisRequestPublishView(BrowserView):
     def _DEFAULT_TEMPLATE(self):
         registry = getUtility(IRegistry)
         return registry.get(
-            'bika.lims.analysisrequest.default_arreport_template', 'default.pt')
+            'bika.lims.analysisrequest.default_arreport_template',
+            'default.pt')
 
     def next_certificate_number(self):
         """Get a new certificate id.  These are throwaway IDs, until the
@@ -270,11 +273,11 @@ class AnalysisRequestPublishView(BrowserView):
         return self.request.form.get('landscape', '0').lower() in ['true', '1']
 
     def localise_images(self, htmlreport):
-        """WeasyPrint will attempt to retrieve attachments directly from the URL
-        referenced in the HTML report, which may refer back to a single-threaded
-        (and currently occupied) zeoclient, hanging it.  All "attachments"
-        using urls ending with at_download/AttachmentFile must be converted
-        to local files.
+        """WeasyPrint will attempt to retrieve attachments directly from
+        the URL referenced in the HTML report, which may refer back to
+        a single-threaded (and currently occupied) zeoclient, hanging it.
+        All "attachments" using urls ending with at_download/AttachmentFile
+        must be converted to local files.
 
         Returns a list of files which were created, and a modified copy
         of htmlreport.
@@ -711,9 +714,10 @@ class AnalysisRequestDigester:
         # dictionaries from the schemas of objects.
         self.SKIP_FIELDNAMES = [
             'allowDiscussion', 'subject', 'location', 'contributors',
-            'creators', 'effectiveDate', 'expirationDate', 'language', 'rights',
-            'relatedItems', 'modification_date', 'immediatelyAddableTypes',
-            'locallyAllowedTypes', 'nextPreviousEnabled', 'constrainTypesMode',
+            'creators', 'effectiveDate', 'expirationDate', 'language',
+            'rights', 'relatedItems', 'modification_date',
+            'immediatelyAddableTypes', 'locallyAllowedTypes',
+            'nextPreviousEnabled', 'constrainTypesMode',
             'RestrictedCategories', 'Digest',
         ]
 
@@ -939,10 +943,10 @@ class AnalysisRequestDigester:
                 'prepublish': False,
                 'child_analysisrequest': None,
                 'parent_analysisrequest': None,
-                'resultsinterpretation':ar.getResultsInterpretation(),
+                'resultsinterpretation': ar.getResultsInterpretation(),
                 'ar_attachments': self._get_ar_attachments(ar),
                 'an_attachments': self._get_an_attachments(ar),
-        }
+                }
 
         # Sub-objects
         excludearuids.append(ar.UID())
@@ -957,15 +961,16 @@ class AnalysisRequestDigester:
 
         wf = ar.portal_workflow
         allowed_states = ['verified', 'published']
-        data['prepublish'] = wf.getInfoFor(ar,
-                                           'review_state') not in allowed_states
+        data['prepublish'] = wf.getInfoFor(
+                ar, 'review_state') not in allowed_states
 
         data['contact'] = self._contact_data(ar)
         data['client'] = self._client_data(ar)
         data['sample'] = self._sample_data(ar)
         data['batch'] = self._batch_data(ar)
         data['specifications'] = self._specs_data(ar)
-        data['analyses'] = self._analyses_data(ar, ['verified', 'published'])
+        data['analyses'] = self._analyses_data(
+                ar, ['verified', 'published'], data['sample'])
         data['qcanalyses'] = self._qcanalyses_data(ar,
                                                    ['verified', 'published'])
         data['points_of_capture'] = sorted(
@@ -977,8 +982,9 @@ class AnalysisRequestDigester:
              an['previous_results']]) > 0
         data['hasblanks'] = len([an['reftype'] for an in data['qcanalyses'] if
                                  an['reftype'] == 'b']) > 0
-        data['hascontrols'] = len([an['reftype'] for an in data['qcanalyses'] if
-                                   an['reftype'] == 'c']) > 0
+        data['hascontrols'] = len(
+                [an['reftype'] for an in data['qcanalyses']
+                    if an['reftype'] == 'c']) > 0
         data['hasduplicates'] = len(
             [an['reftype'] for an in data['qcanalyses'] if
              an['reftype'] == 'd']) > 0
@@ -1173,17 +1179,21 @@ class AnalysisRequestDigester:
             physical_address = _format_address(
                 contact.getPhysicalAddress()) if contact else ''
             postal_address =\
-                    _format_address(contact.getPostalAddress())\
+                _format_address(contact.getPostalAddress())\
                 if contact else ''
             data = {'id': member.id,
                     'fullname': to_utf8(cfullname) if cfullname else to_utf8(
                         mfullname),
                     'email': cemail if cemail else memail,
-                    'business_phone': contact.getBusinessPhone() if contact else '',
-                    'business_fax': contact.getBusinessFax() if contact else '',
+                    'business_phone':
+                        contact.getBusinessPhone() if contact else '',
+                    'business_fax':
+                        contact.getBusinessFax() if contact else '',
                     'home_phone': contact.getHomePhone() if contact else '',
-                    'mobile_phone': contact.getMobilePhone() if contact else '',
-                    'job_title': to_utf8(contact.getJobTitle()) if contact else '',
+                    'mobile_phone':
+                        contact.getMobilePhone() if contact else '',
+                    'job_title':
+                        to_utf8(contact.getJobTitle()) if contact else '',
                     'physical_address': physical_address,
                     'postal_address': postal_address,
                     'home_page': to_utf8(mhomepage)}
@@ -1270,7 +1280,7 @@ class AnalysisRequestDigester:
 
         return data
 
-    def _analyses_data(self, ar, analysis_states=None):
+    def _analyses_data(self, ar, analysis_states=None, sample=None):
         if not analysis_states:
             analysis_states = ['verified', 'published']
         analyses = []
@@ -1290,7 +1300,7 @@ class AnalysisRequestDigester:
                 continue
 
             # Build the analysis-specific dict
-            andict = self._analysis_data(an, dm)
+            andict = self._analysis_data(an, dm, sample)
 
             # Are there previous results for the same AS and batch?
             andict['previous'] = []
@@ -1312,10 +1322,18 @@ class AnalysisRequestDigester:
                 andict['previous_results'] = ", ".join(
                     [p['formatted_result'] for p in andict['previous'][-5:]])
 
-            analyses.append(andict)
+            # Append addition analysis dicts for each unit conversion
+            if andict.get('unit_conversions', False):
+                for uc_uid in andict['unit_conversions']:
+                    analyses.append(_convert_unit_of(uc_uid, andict, an))
+
+            # Append primary analysis if appropriate
+            if not andict.get('hide_primary_result'):
+                analyses.append(andict)
+
         return analyses
 
-    def _analysis_data(self, analysis, decimalmark=None):
+    def _analysis_data(self, analysis, decimalmark=None, sample=None):
 
         andict = {'obj': analysis,
                   'id': analysis.id,
@@ -1323,8 +1341,8 @@ class AnalysisRequestDigester:
                   'keyword': analysis.getKeyword(),
                   'scientific_name': analysis.getScientificName(),
                   'accredited': analysis.getAccredited(),
-                  'point_of_capture': to_utf8(
-                      POINTS_OF_CAPTURE.getValue(analysis.getPointOfCapture())),
+                  'point_of_capture': to_utf8(POINTS_OF_CAPTURE.getValue(
+                      analysis.getPointOfCapture())),
                   'category': to_utf8(analysis.getCategoryTitle()),
                   'result': analysis.getResult(),
                   'isnumber': isnumber(analysis.getResult()),
@@ -1340,12 +1358,15 @@ class AnalysisRequestDigester:
                   'resultdm': to_utf8(analysis.getResultDM()),
                   'outofrange': False,
                   'type': analysis.portal_type,
-                  'reftype': analysis.getReferenceType() \
-                      if hasattr(analysis, 'getReferenceType') \
-                      else None,
+                  'reftype': analysis.getReferenceType()
+                  if hasattr(analysis, 'getReferenceType')
+                  else None,
                   'worksheet': None,
                   'specs': {},
-                  'formatted_specs': ''}
+                  'formatted_specs': '',
+                  'unit_conversions': [],
+                  'hide_primary_result': False,
+                  }
 
         if analysis.portal_type == 'DuplicateAnalysis':
             andict['reftype'] = 'd'
@@ -1356,7 +1377,8 @@ class AnalysisRequestDigester:
             if ws and len(ws) > 0 else None
         andict['refsample'] = analysis.getSample().id \
             if analysis.portal_type == 'Analysis' \
-            else '%s - %s' % (analysis.aq_parent.id, analysis.aq_parent.Title())
+            else '%s - %s' % (
+                    analysis.aq_parent.id, analysis.aq_parent.Title())
 
         if analysis.portal_type == 'ReferenceAnalysis':
             # The analysis is a Control or Blank. We might use the
@@ -1380,8 +1402,8 @@ class AnalysisRequestDigester:
         # expressed in sci notation, that may include <sup></sup> html tags.
         # Please note the default value for the 'html' parameter from
         # getFormattedResult signature is set to True, so the service will
-        # already take into account LDLs and UDLs symbols '<' and '>' and escape
-        # them if necessary.
+        # already take into account LDLs and UDLs symbols '<' and '>'
+        # and escape# them if necessary.
         andict['formatted_result'] = fresult
 
         fs = ''
@@ -1404,6 +1426,25 @@ class AnalysisRequestDigester:
                 if ret and ret['out_of_range']:
                     andict['outofrange'] = True
                     break
+
+        # Get unit conversions for result
+        st_uid = None
+        if sample and sample.get('sample_type') and \
+           sample['sample_type'].get('obj'):
+            st_uid = sample['sample_type']['obj'].UID()
+        if st_uid:
+            for unit_conversion in analysis.getUnitConversions():
+                if unit_conversion.get('Unit') and \
+                   unit_conversion.get('SampleType', None) == st_uid:
+                    # Append new unit conversion
+                    andict['unit_conversions'].append(unit_conversion['Unit'])
+
+                    # if any one UC record indicates hiding, hide it
+                    hide_primary = \
+                        unit_conversion.get('HidePrimaryUnit', '0') == '1'
+                    if hide_primary:
+                        andict['hide_primary_result'] = True
+
         return andict
 
     def _qcanalyses_data(self, ar, analysis_states=None):
@@ -1570,3 +1611,28 @@ def _format_address(address):
         addr = ''.join(["<span>%s</span>" % address.get(v) for v in keys
                         if address.get(v, None)])
     return "<div class='address'>%s</div>" % addr
+
+def _convert_unit_of(uc_uid, andict, context=None):
+    """
+    Duplicate an existing analysis dict and add the converted values
+    according to the given Unit Convsersion record
+    """
+    new = dict(andict)
+    unit_conversion = api.get_object_by_uid(uid=uc_uid)
+    new['unit'] = unit_conversion.converted_unit
+    new['formatted_unit'] = unit_conversion.converted_unit
+    if andict.get('result'):
+        new['formatted_result'] = new['result'] = \
+            convert_unit(
+                    andict['result'], 
+                    unit_conversion.formula,
+                    andict['obj'].getPrecision())
+    return new
+
+def _get_unit_conversion(unit_conversion, sample_type_id):
+    if unit_conversion.get('Unit') and \
+       unit_conversion.get('SampleType', None) == sample_type_id:
+        andict['unit_conversions'].append(unit_conversion['Unit'])
+        if andict['hide_primary_result'] == False:
+            andict['hide_primary_result'] = \
+                unit_conversion.get('HidePrimaryUnit', '0') == '1'