+++ bika/lims/browser/js/bika.lims.utils.calcs.js
@@ -195,7 +195,13 @@ function CalculationUtils() {
                     // put result values in their boxes
                     for(i=0;i<$(data['results']).length;i++){
                         result = $(data['results'])[i];
-                         $("input[uid='"+result.uid+"']").filter("input[field='Result']").val(result.result);
+                        $("input[uid='"+result.uid+"']").filter("input[field='Result']").val(result.result);
+
+                        var converted_result_input = $("input[uid='"+result.uid+"']").filter("input[field='ConvertedResult']");
+                        converted_result_input.val(result.converted_result);
+                        converted_result_input.prop('type', 'text');
+                        converted_result_input.prop('readonly', true);
+                        converted_result_input.prop('style', 'font-size: 100%');
 
                         $('[type="hidden"]').filter("[field='ResultDM']").filter("[uid='"+result.uid+"']").val(result.dry_result);
                         $($('[type="hidden"]').filter("[field='ResultDM']").filter("[uid='"+result.uid+"']").siblings()[0]).empty().append(result.dry_result);