+++ bika/lims/content/analysis.py
@@ -11,7 +11,7 @@ from bika.lims import PROJECTNAME
 from bika.lims.content.abstractroutineanalysis import AbstractRoutineAnalysis
 from bika.lims.content.abstractroutineanalysis import schema
 from bika.lims.interfaces import IRoutineAnalysis, ISamplePrepWorkflow
-from bika.lims.workflow import getCurrentState, in_state
+from bika.lims.workflow import in_state
 from bika.lims.workflow.analysis import STATE_RETRACTED, STATE_REJECTED
 from zope.interface import implements
 
@@ -62,6 +62,16 @@ class Analysis(AbstractRoutineAnalysis):
             siblings.append(sibling)
 
         return siblings
+
+    @security.public
+    def workflow_script_publish(self):
+        """
+        If this is not here, acquisition causes recursion into
+        AR workflow_script_publish method and, if there are enough
+        analyses, it will result in "RuntimeError: maximum recursion
+        depth exceeded"
+        """
+        pass
 
 
 registerType(Analysis, PROJECTNAME)