+++ bika/lims/utils/analysisrequest.py
@@ -5,8 +5,10 @@
 # Copyright 2018 by it's authors.
 # Some rights reserved. See LICENSE.rst, CONTRIBUTORS.rst.
 
-from Products.CMFCore.utils import getToolByName
-from Products.CMFPlone.utils import safe_unicode
+import json
+import os
+import tempfile
+from bika.lims import api
 from bika.lims import bikaMessageFactory as _
 from bika.lims import logger
 from bika.lims.idserver import renameAfterCreation
@@ -21,14 +23,19 @@ from bika.lims.utils.samplepartition import create_samplepartition
 from bika.lims.workflow import doActionFor
 from bika.lims.workflow import doActionsFor
 from bika.lims.workflow import getReviewHistoryActionsList
+from collective.taskqueue.interfaces import ITaskQueue
 from copy import deepcopy
 from email.mime.multipart import MIMEMultipart
 from email.mime.text import MIMEText
 from email.Utils import formataddr
-from plone import api
+from plone import api as ploneapi
+from Products.CMFCore.utils import getToolByName
+from Products.CMFPlone.utils import safe_unicode
 from Products.CMFPlone.utils import _createObjectByType
-import os
-import tempfile
+from Products.Five.browser import BrowserView
+from smtplib import SMTPRecipientsRefused
+from zope.component import queryUtility
+
 
 def create_analysisrequest(client, request, values, analyses=None,
                            partitions=None, specifications=None, prices=None):
@@ -163,6 +170,11 @@ def create_analysisrequest(client, request, values, analyses=None,
     if reject_field and reject_field.get('checkbox', False):
         doActionFor(ar, 'reject')
 
+    bika_setup = api.get_bika_setup()
+    if bika_setup.getRegisterAsReceivd():
+        # Transition AR and analyses to Received state
+        doActionFor(ar, 'receive')
+
     return ar
 
 
@@ -250,7 +262,6 @@ def _resolve_items_to_service_uids(items):
         If an item that doesn't match any of the criterias above is found, the
         function will raise a RuntimeError
     """
-    portal = None
     bsc = None
     service_uids = []
 
@@ -274,8 +285,7 @@ def _resolve_items_to_service_uids(items):
             continue
 
         # Maybe object UID.
-        portal = portal if portal else api.portal.get()
-        bsc = bsc if bsc else getToolByName(portal, 'bika_setup_catalog')
+        bsc = api.get_tool('bika_setup_catalog')
         brains = bsc(UID=item)
         if brains:
             uid = brains[0].UID
@@ -321,7 +331,8 @@ def notify_rejection(analysisrequest):
 
     # This is the template to render for the pdf that will be either attached
     # to the email and attached the the Analysis Request for further access
-    tpl = AnalysisRequestRejectPdfView(analysisrequest, analysisrequest.REQUEST)
+    tpl = AnalysisRequestRejectPdfView(
+            analysisrequest, analysisrequest.REQUEST)
     html = tpl.template()
     html = safe_unicode(html).encode('utf-8')
     filename = '%s-rejected' % arid
@@ -354,7 +365,8 @@ def notify_rejection(analysisrequest):
     # compose and send email.
     mailto = []
     lab = analysisrequest.bika_setup.laboratory
-    mailfrom = formataddr((encode_header(lab.getName()), lab.getEmailAddress()))
+    mailfrom = formataddr(
+            (encode_header(lab.getName()), lab.getEmailAddress()))
     mailsubject = _('%s has been rejected') % arid
     contacts = [analysisrequest.getContact()] + analysisrequest.getCCContact()
     for contact in contacts:
@@ -378,7 +390,93 @@ def notify_rejection(analysisrequest):
         host = getToolByName(analysisrequest, 'MailHost')
         host.send(mime_msg.as_string(), immediate=True)
     except:
-        logger.warning(
-            "Email with subject %s was not sent (SMTP connection error)" % mailsubject)
+        logger.warning("Email with subject {} was not sent"
+                       " (SMTP connection error)".format(mailsubject))
 
     return True
+
+
+class Async_AR_Utils(BrowserView):
+
+    def async_create_analysisrequest(self):
+        msgs = []
+        form = self.request.form
+        records = json.loads(form.get('records', '[]'))
+        attachments = json.loads(form.get('attachments', '[]'))
+        ARs = []
+        logger.info('Async create %s records' % len(records))
+        for n, record in enumerate(records):
+            client_uid = record.get("Client")
+            client = api.get_object_by_uid(client_uid)
+
+            if not client:
+                msgs.append("Error: Client {} found".format(client_uid))
+                continue
+
+            # get the specifications and pass them directly to the AR
+            # create function.
+            specifications = record.pop("Specifications", {})
+
+            # Create the Analysis Request
+            ar = create_analysisrequest(
+                client,
+                self.request,
+                values=record,
+                specifications=specifications)
+            ARs.append(ar.getId())
+
+            _attachments = []
+            for attachment in attachments.get(n, []):
+                if not attachment.filename:
+                    continue
+                att = _createObjectByType("Attachment", self.context, tmpID())
+                att.setAttachmentFile(attachment)
+                att.processForm()
+                _attachments.append(att)
+            if _attachments:
+                ar.setAttachment(_attachments)
+
+        if len(ARs) == 1:
+            msgs.append('Created AR {}'.format(ARs[0]))
+        elif len(ARs) > 1:
+            msgs.append('Created ARs {}'.format(', '.join(ARs)))
+        else:
+            msgs.append('No ARs created')
+        message = '; '.join(msgs)
+        logger.info('AR Creation complete: {}'.format(message))
+        self._email_analyst(message)
+        return
+
+    def queue_count(self):
+        task_queue = queryUtility(ITaskQueue, name='ar-create')
+        if task_queue is None:
+            return 0
+        return 'Len = %s' % len(task_queue)
+
+    def _email_analyst(self, message):
+        mail_template = """
+Dear {name},
+
+Analysis request creation completed with the following messages:
+{message}
+
+Cheers
+Bika LIMS
+"""
+        member = ploneapi.user.get_current()
+        mail_host = ploneapi.portal.get_tool(name='MailHost')
+        from_email = mail_host.email_from_address
+        to_email = member.getProperty('email')
+        subject = 'AR Creation Complete'
+        if len(to_email) == 0:
+            to_email = from_email
+            subject = 'AR Creation by admin user is complete'
+        mail_text = mail_template.format(
+                        name=member.getProperty('fullname'),
+                        message=message)
+        try:
+            return mail_host.send(
+                        mail_text, to_email, from_email,
+                        subject=subject, charset="utf-8", immediate=False)
+        except SMTPRecipientsRefused:
+            raise SMTPRecipientsRefused('Recipient address rejected by server')